Move bins_sum, bins_mean into bins_reduction
