+++ lib/cmake/scipp-functions.cmake
@@ -82,9 +82,9 @@ scipp_function("inplace" logical operator&= OP logical_and_equals SKIP_PYTHON)
 scipp_function("inplace" logical operator^= OP logical_xor_equals SKIP_PYTHON)
 setup_scipp_category(logical)
 
-scipp_function("unary" bins bin_sizes SKIP_VARIABLE BASE_INCLUDE dataset/bins.h)
-scipp_function("unary" bins bins_sum SKIP_VARIABLE BASE_INCLUDE dataset/bins.h)
-scipp_function("unary" bins bins_mean SKIP_VARIABLE BASE_INCLUDE dataset/bins.h)
+scipp_function("unary" bins bin_sizes SKIP_VARIABLE BASE_INCLUDE variable/bins.h)
+scipp_function("unary" bins bins_sum SKIP_VARIABLE BASE_INCLUDE dataset/bins_reduction.h)
+scipp_function("unary" bins bins_mean SKIP_VARIABLE BASE_INCLUDE dataset/bins_reduction.h)
 setup_scipp_category(bins)
 
 scipp_function("reduction" reduction sum SKIP_VARIABLE)
